moves router into bloomcompactor Signed-off-by: Owen Diehl <ow.diehl@gmail.com>
